backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
alembic==1.12.0
SQLAlchemy==2.0.21
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
blinker==1.6.3
google-generativeai==0.3.0
requests==2.31.0
